+++ crates/ryo3-jiff/src/ry_time.rs
@@ -52,6 +52,11 @@ impl RyTime {
     // ========================================================================
     // CLASS ATTRS
     // ========================================================================
+    #[classattr]
+    fn __match_args__() -> (&'static str, &'static str, &'static str, &'static str) {
+        ("hour", "minute", "second", "subsec_nanosecond")
+    }
+
     #[expect(non_snake_case)]
     #[classattr]
     fn MIN() -> Self {